Resolve problem with __file__
